ci: gate releases on production audits
